\name{lrpackage-package}
\alias{lrpackage-package}
\alias{lrpackage}
\docType{package}
\title{
\packageTitle{lrpackage}
}
\description{
\packageDescription{lrpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lrpackage}
\packageIndices{lrpackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{lrpackage}

Maintainer: \packageMaintainer{lrpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
